Implement email sending feature

+++ src/main/java/com/kashkina/portfolio/dto/skills/TechStackCategoryDTO.java
@@ -9,12 +9,12 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 public class TechStackCategoryDTO {
-    private Long id;
+    private Integer id;
     private String name;
     private Integer orderIndex;
     private List<TechStackItemDTO> items;
 
-    public TechStackCategoryDTO(Long id, String name, Integer orderIndex, List<TechStackItemDTO> items) {
+    public TechStackCategoryDTO(Integer id, String name, Integer orderIndex, List<TechStackItemDTO> items) {
         this.id = id;
         this.name = name;
         this.orderIndex = orderIndex;
