+++ src/main/java/com/kashkina/portfolio/repository/projects/ProjectRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<Project, Long> {
+public interface ProjectRepository extends JpaRepository<Project, Integer> {
 
     @Override
     @EntityGraph(attributePaths = {"technologies", "features", "screenshots"})
