+++ src/main/java/com/kashkina/portfolio/repository/skills/SoftSkillRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SoftSkillRepository extends JpaRepository<SoftSkill, Long> {}
+public interface SoftSkillRepository extends JpaRepository<SoftSkill, Integer> {}
