+++ src/main/java/com/kashkina/portfolio/repository/skills/TechStackCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TechStackCategoryRepository extends JpaRepository<TechStackCategory, Long> {}
+public interface TechStackCategoryRepository extends JpaRepository<TechStackCategory, Integer> {}
